# docs/v2/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "AI Agent Productivity Platform API V2",
        "contact": {},
        "version": "v0.0.1"
    },
    "basePath": "/v2",
    "paths": {
...
        "/statistic/app/chart": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "调用趋势 + 智能体/工作流/对话流/知识问答排行",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app_observability.statistic"
                ],
                "summary": "获取应用统计趋势+排行",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AppStatisticV2ChartReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.AppStatisticV2Chart"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/statistic/app/list": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "按应用聚合的调用统计（分页）。sortField 可选数值指标：callCount、callFailure、streamCount、nonStreamCount、failureRate、avgFirstTokenLatency、avgCosts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app_observability.statistic"
                ],
                "summary": "获取应用调用统计列表",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AppStatisticV2ListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/response.PageResult"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/response.AppStatisticV2ListItem"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/statistic/app/list/export": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "导出应用调用统计列表。sortField 可选数值指标：callCount、callFailure、streamCount、nonStreamCount、streamFailure、nonStreamFailure、failureRate、avgFirstTokenLatency、avgCosts、firstTokenLatency、costs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "app_observability.statistic"
                ],
                "summary": "导出应用调用统计列表",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AppStatisticV2ExportListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/statistic/app/list/model": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "指定应用下的模型使用统计（分页）。钻取需传当前主表行的 source/appId/moduleCreatorUserId/moduleCreatorOrgId。sortField 可选数值指标：totalTokens、promptTokens、completionTokens、callCount、callFailure、failureRate、avgFirstTokenLatency、avgCosts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app_observability.statistic"
                ],
                "summary": "获取应用模型使用列表",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AppStatisticV2ModelListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/response.PageResult"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/response.AppStatisticV2ModelListItem"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/statistic/app/list/model/export": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "导出应用模型使用列表。sortField 可选数值指标：totalTokens、promptTokens、completionTokens、callCount、callFailure、streamCount、nonStreamCount、streamFailure、nonStreamFailure、failureRate、avgFirstTokenLatency、avgCosts、firstTokenLatency、costs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "app_observability.statistic"
                ],
                "summary": "导出应用模型使用列表",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AppStatisticV2ModelExportReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/statistic/app/list/user": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "指定应用下的用户使用统计（分页）。钻取需传当前主表行的 source/appId/moduleCreatorUserId/moduleCreatorOrgId。sortField 可选数值指标：callCount、callFailure、streamCount、nonStreamCount、failureRate、avgFirstTokenLatency、avgCosts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app_observability.statistic"
                ],
                "summary": "获取应用用户使用列表",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AppStatisticV2UserListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/response.PageResult"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/response.AppStatisticV2UserListItem"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/statistic/app/list/user/export": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "导出应用用户使用列表。sortField 可选数值指标：callCount、callFailure、streamCount、nonStreamCount、streamFailure、nonStreamFailure、failureRate、avgFirstTokenLatency、avgCosts、firstTokenLatency、costs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "app_observability.statistic"
                ],
                "summary": "导出应用用户使用列表",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AppStatisticV2UserExportReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/statistic/app/overview": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "10张指标卡（调用次数、Tokens、流式/非流式维度）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app_observability.statistic"
                ],
                "summary": "获取应用统计概览",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AppStatisticV2Req"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.AppStatisticV2Overview"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/statistic/app/record/export": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "导出应用调用明细（不支持用户排序；后端固定按调用时间倒序）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "app_observability.statistic"
                ],
                "summary": "导出应用调用明细",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AppStatisticV2RecordExportReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/statistic/app/record/list": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "应用调用明细记录（分页，不支持用户排序；后端固定按调用时间倒序）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app_observability.statistic"
                ],
                "summary": "获取应用调用明细列表",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AppStatisticV2RecordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/response.PageResult"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/response.AppStatisticV2RecordItem"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/statistic/app/select": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "viewScope=published 查主表（我发布的）；used 查聚合表（我使用的）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app_observability.statistic"
                ],
                "summary": "获取应用统计下拉列表",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AppStatisticV2SelectReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/response.ListResult"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/response.MyAppItem"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.AppStatisticV2ChartReq": {
            "type": "object",
            "required": [
                "endDate",
                "startDate",
                "viewScope"
            ],
            "properties": {
                "apps": {
                    "description": "appId 列表，空=全部",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "endDate": {
                    "type": "string"
                },
                "limit": {
                    "description": "排行 TopN，默认 5",
                    "type": "integer"
                },
                "module": {
                    "description": "板块 wga|skill|knowledge|...，空=全部",
                    "type": "string"
                },
                "orgIds": {
                    "description": "空=按角色解析（见上方约定）；[\"ALL\"]=可见全部组织；[\"id\",...]=指定组织",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "source": {
                    "description": "调用来源 web|openapi|webURL；空=全部（钻取见子结构必填）",
                    "type": "string"
                },
                "startDate": {
                    "description": "yyyy-MM-dd",
                    "type": "string"
                },
                "userIds": {
                    "description": "空=按角色解析（见上方约定）；[\"ALL\"]=已解析组织下全部用户；[\"id\",...]=指定用户",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "viewScope": {
                    "description": "published|used",
                    "type": "string"
                }
            }
        },
...
        "response.AppStatisticV2Chart": {
            "type": "object",
            "properties": {
                "rank": {
                    "$ref": "#/definitions/response.AppStatisticV2Rank"
                },
                "trend": {
                    "$ref": "#/definitions/response.AppStatisticV2Trend"
                }
            }
        },
...
        "response.ListResult": {
            "type": "object",
            "properties": {
                "list": {},
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}